initial revision

+++ support/ab.c
@@ -353,7 +353,7 @@ apr_size_t reqlen;
 char buffer[8192];
 
 /* interesting percentiles */
-int percs[] = {50, 66, 75, 80, 90, 95, 98, 99, 99.9, 99.99, 100};
+float percs[] = {50, 66, 75, 80, 90, 95, 98, 99, 99.9, 99.99, 100};
 
 struct connection *con;     /* connection array */
 struct data *stats;         /* data for each request */
@@ -956,7 +956,7 @@ static void output_results(int sig)
                     printf(" 100%%  %5" APR_TIME_T_FMT " (longest request)\n",
                            ap_round_ms(stats[done - 1].time));
                 else
-                    printf("  %d%%  %5" APR_TIME_T_FMT "\n", percs[i],
+                    printf("  %f%%  %5" APR_TIME_T_FMT "\n", percs[i],
                            ap_round_ms(stats[(int) (done * percs[i] / 100)].time));
             }
         }
